+++ src/spdx_tools/spdx3/validation/rdf/model.ttl
