+++ src/spdx_tools/spdx3/validation/rdf/shacl_validation.py
@@ -10,11 +10,11 @@ def validate_against_shacl_from_file(
     data_file: str, shacl_file: str, data_format: Optional[str] = "json-ld", shacl_format: Optional[str] = "ttl"
 ):
     data_graph = Graph()
-    with open(data_file) as file:
+    with open(data_file, "rb") as file:
         data_graph.parse(file, format=data_format)
 
     shacl_graph = Graph()
-    with open(shacl_file) as file:
+    with open(shacl_file, "rb") as file:
         shacl_graph.parse(file, format=shacl_format)
 
     return validate(data_graph=data_graph, shacl_graph=shacl_graph, ont_graph=shacl_graph)
