+++ tests/spdx3/validation/json_ld/test_shacl_validation.py
@@ -5,7 +5,7 @@ import os
 
 import pytest
 
-from spdx_tools.spdx3.validation.json_ld.shacl_validation import validate_against_shacl_from_file
+from spdx_tools.spdx3.validation.rdf.shacl_validation import validate_against_shacl_from_file
 
 
 @pytest.mark.skip("Currently the validation against SHACL fails, refer to process.md and the known limitations.")
@@ -14,7 +14,7 @@ def test_shacl_validation():
     conforms, results_graph, results_text = validate_against_shacl_from_file(
         data_file=os.path.join(os.path.dirname(__file__), "../../../SPDX3_jsonld_test.jsonld"),
         shacl_file=os.path.join(
-            os.path.dirname(__file__), "../../../../src/spdx_tools/spdx3/writer/json_ld/model.ttl"
+            os.path.dirname(__file__), "../../../../src/spdx_tools/spdx3/validation/rdf/model.ttl"
         ),
     )
     # results_graph.serialize("validation_result.rdf.xml", format="pretty-xml")
